pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "neuropose"
version = "0.1.0.dev0"
description = "3D human pose estimation pipeline for clinical movement research, built on MeTRAbs."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11,<3.12"
authors = [
  { name = "Liqi Shu", email = "liqi_shu@brown.edu" },
  { name = "Levi Neuwirth", email = "ln@levineuwirth.org" },
  { name = "David Man", email = "david_man@brown.edu" },
  { name = "Praneeth Tummala", email = "praneeth_tummala@brown.edu" },
]
maintainers = [
  { name = "Levi Neuwirth", email = "ln@levineuwirth.org" },
]
keywords = [
  "pose-estimation",
  "biomechanics",
  "motion-capture",
  "clinical-research",
  "metrabs",
]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Image Recognition",
  "Topic :: Scientific/Engineering :: Medical Science Apps.",
  "Typing :: Typed",
]

# Runtime dependencies. Versions intentionally unpinned at this stage: the
# TensorFlow / MeTRAbs stack is pending a sanity check against the updated
# upstream MeTRAbs release. Tightened pins will land alongside the MeTRAbs
# model loader and smoke test.
dependencies = [
  "typer>=0.12",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "pyyaml>=6.0",
  "numpy>=1.26",
  "opencv-python-headless>=4.9",
  "matplotlib>=3.8",
  "tensorflow>=2.15,<3.0",
  "tensorflow-hub>=0.16",
]

[project.optional-dependencies]
analysis = [
  "fastdtw>=0.3.4",
  "scipy>=1.12",
  "scikit-learn>=1.4",
  "sktime>=0.28",
]

[project.urls]
Homepage = "https://git.levineuwirth.org/neuwirth/neuropose"
Repository = "https://git.levineuwirth.org/neuwirth/neuropose"
Issues = "https://git.levineuwirth.org/neuwirth/neuropose/issues"

# CLI entry point will be wired up in commit 7 once src/neuropose/cli.py lands.
# [project.scripts]
# neuropose = "neuropose.cli:app"

# ---------------------------------------------------------------------------
# Dependency groups (PEP 735). Install a group with `uv sync --group dev` or
# `uv pip install --group dev`. These are not shipped with the wheel.
# ---------------------------------------------------------------------------
[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.8",
  "pyright>=1.1.390",
  "pre-commit>=4.0",
]

# ---------------------------------------------------------------------------
# Ruff (linter + formatter). See https://docs.astral.sh/ruff/rules/ for the
# meaning of each rule code. The selection is deliberately broader than the
# Python community default — we want lint noise early rather than cruft late.
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
extend-exclude = ["notebooks"]

[tool.ruff.lint]
select = [
  "E", "W",   # pycodestyle
  "F",        # pyflakes
  "I",        # isort
  "B",        # flake8-bugbear
  "UP",       # pyupgrade
  "SIM",      # flake8-simplify
  "RUF",      # ruff-specific
  "N",        # pep8-naming
  "C4",       # flake8-comprehensions
  "PTH",      # flake8-use-pathlib
  "PT",       # flake8-pytest-style
  "TID",      # flake8-tidy-imports
  "NPY",      # numpy-specific
  "D",        # pydocstyle
]
ignore = [
  "D100",  # missing module docstring — too strict for internal modules
  "D104",  # missing package docstring — too strict for __init__.py
  "D203",  # conflicts with D211
  "D213",  # conflicts with D212
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D", "PLR2004"]
"scripts/**" = ["D"]
"src/neuropose/_*.py" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
known-first-party = ["neuropose"]

[tool.ruff.format]
docstring-code-format = true

# ---------------------------------------------------------------------------
# Pytest.
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ["--strict-markers", "--strict-config", "-ra"]
markers = [
  "slow: tests that require model download or are otherwise slow (deselect with '-m \"not slow\"')",
]

# ---------------------------------------------------------------------------
# Pyright. "standard" rather than "strict" because the TensorFlow / OpenCV /
# scikit-learn stubs would generate thousands of unknown-type warnings under
# strict mode. We tighten toward strict after the MeTRAbs stack is pinned.
# ---------------------------------------------------------------------------
[tool.pyright]
include = ["src", "tests"]
exclude = [
  "**/__pycache__",
  "**/.venv",
  "notebooks",
]
pythonVersion = "3.11"
typeCheckingMode = "standard"
reportMissingTypeStubs = "none"
reportUnknownMemberType = "none"
reportUnknownArgumentType = "none"
reportUnknownVariableType = "none"
reportUnknownParameterType = "none"

# ---------------------------------------------------------------------------
# Hatch build configuration.
# ---------------------------------------------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["src/neuropose"]

[tool.hatch.build.targets.sdist]
include = [
  "src/neuropose",
  "README.md",
  "LICENSE",
  "AUTHORS.md",
  "CITATION.cff",
]
